% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehelp-example.R
\name{eexample}
\alias{eexample}
\title{function that allows to execute the examples from user defined functions}
\usage{
eexample(..., skip.donts = FALSE)
}
\arguments{
\item{...}{function name of a user defined fn}

\item{skip.donts}{boolean argument to specify whether dontest or dontrun examples should be skiped or not}
}
\description{
function that allows to execute the examples from user defined functions
}
