% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehelp-defns.R
\name{ehelp.palette}
\alias{ehelp.palette}
\title{This function defines the "palettes" of colours to be used by eHelp()
There are two options right now: "color" and "bw"; default is set to "color"}
\usage{
ehelp.palette(palette = "color")
}
\arguments{
\item{palette}{allows to select which color palette to use}
}
\description{
This function defines the "palettes" of colours to be used by eHelp()
There are two options right now: "color" and "bw"; default is set to "color"
}
\keyword{internal}
