% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehelp-defns.R
\name{format.defns}
\alias{format.defns}
\title{Function to define specifics structures, symbols and 'codes'}
\usage{
\method{format}{defns}(fmt, filename = "", ending = NA)
}
\arguments{
\item{fmt}{an specific file format; possible values are: txt, html, latex, markdown}

\item{ending}{optional argument for selecting when continuing saving into the file}
}
\description{
Function to define specifics structures, symbols and 'codes'
}
\keyword{internal}
