% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehelp.R
\name{help}
\alias{help}
\title{Wrapper Help Function}
\usage{
help(
  topic,
  package = NULL,
  lib.loc = NULL,
  verbose = getOption("verbose"),
  try.all.packages = getOption("help.try.all.packages"),
  help_type = getOption("help_type")
)
}
\arguments{
\item{topic}{topic/or/function name to search for}

\item{package}{package where to search}

\item{lib.loc}{location of R libraries}

\item{verbose}{for diplaying the filename}

\item{try.all.packages}{attempt to go trough all installed packages}

\item{help_type}{format of the displayed help (text,html, or pdf)}
}
\description{
This function is a wrapper around the R's system help() function.
It allows the user to include docstring styles documentation and
displayed it as help or information to the users using the help()
command.
}
\details{
Parameters are the same as in utils::help, see help(help,package='utils') for further details.
}
