% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehelp-obj.R
\name{print.ehelp}
\alias{print.ehelp}
\title{function associated to the ehelp object method}
\usage{
\method{print}{ehelp}(ehelp.obj)
}
\arguments{
\item{ehelp.obj}{ehelp object
keywords internal}
}
\description{
function associated to the ehelp object method
}
