% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehelp-utils.R
\name{print}
\alias{print}
\title{method associated with ehelp objects}
\usage{
print(ehelp.obj)
}
\arguments{
\item{ehelp.obj}{ehelp object}
}
\description{
method associated with ehelp objects
}
\keyword{internal}
