% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehelp-utils.R
\name{processOutput}
\alias{processOutput}
\title{function to process the type output requested}
\usage{
processOutput(ehelp.obj, fnName, fnCorpus, output)
}
\arguments{
\item{ehelp.obj}{ehelp object generated by the ehelp() fn}

\item{fnName}{name of the function}

\item{fnCorpus}{function definition}

\item{output}{type of output, valid options are "txt", "TXT", "html", "HTML", "latex", "LATEX", "markdown", "MARKDOWN"}
}
\description{
function to process the type output requested
}
\keyword{internal}
