% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulatePkg.R
\name{simulatePackage}
\alias{simulatePackage}
\title{function that allows to load the functions from a package in preparation for
CRAN, as if it is being loaded by loading all the fns defined in the R
sub-directory of the package, ie. "myPckg/R"}
\usage{
simulatePackage(pkgLocation = NULL)
}
\arguments{
\item{pkgLocation}{path to the base loaction of the package, under which
is expected to found the R sub-directory}
}
\description{
function that allows to load the functions from a package in preparation for
CRAN, as if it is being loaded by loading all the fns defined in the R
sub-directory of the package, ie. "myPckg/R"
}
