% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehelp-utils.R
\name{write.ehelp}
\alias{write.ehelp}
\title{function for writing ehelp output to a local text file}
\usage{
write.ehelp(X.obj, filename)
}
\arguments{
\item{X.obj}{ehelp object}

\item{filename}{name of the file where to write the documentation}
}
\description{
function for writing ehelp output to a local text file
}
\keyword{internal}
