% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehelp-utils.R
\name{write.fmt}
\alias{write.fmt}
\title{function for writing ehelp output to a local LaTeX file}
\usage{
write.fmt(X.obj, format, filename, leaveOpen = FALSE)
}
\arguments{
\item{X.obj}{ehelp object}

\item{format}{type of output: txt, latex, html}

\item{filename}{name of the file where to write the documentation}

\item{leaveOpen}{boolean argument to leave the file open}
}
\description{
function for writing ehelp output to a local LaTeX file
}
\keyword{internal}
