% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehelp-utils.R
\name{write.Fncorpus}
\alias{write.Fncorpus}
\title{Function to write into a file the listing of the fn.}
\usage{
write.Fncorpus(
  fnName,
  fnListing,
  filename,
  lines = paste0("/* ################################################ */"),
  begining = "",
  ending = "",
  EoL = "\\r\\n"
)
}
\arguments{
\item{fnName}{name of the function}

\item{fnListing}{listing of fn to write to file}

\item{filename}{name of the file where to write the documentation}

\item{lines}{optional argument to specify the separation lines to be used}

\item{begining}{an optional argument to specify the initial lines of HTML code}

\item{ending}{an optional argument to specify the last lines of HTML code}

\item{EoL}{optional argument to specify the EOL (if need, eg. for html)}
}
\description{
Function to write into a file the listing of the fn.
}
\keyword{internal}
