% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehelp-utils.R
\name{write.Info}
\alias{write.Info}
\title{function for including informaton about time and date in the generated files}
\usage{
write.Info(
  filename,
  lines = paste0("-----------------------------------------------"),
  pre = "",
  post = "",
  EoL = ""
)
}
\arguments{
\item{filename}{name of the file where to write the documentation}

\item{lines}{optional argument to specify the separation lines to be used}

\item{post}{optional argument to specify post-information code to be added (if need, eg. for html)}

\item{EoL}{optional argument to specify the EOL (if need, eg. for html)}
}
\description{
function for including informaton about time and date in the generated files
}
\keyword{internal}
