% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehelp.R
\name{ehelp}
\alias{ehelp}
\title{Enhanced-Help Function: ehelp()
This function displays docstring style comments used as help liners for user
defined functions.}
\usage{
ehelp(fun, fn.name = as.character(substitute(fun)))
}
\arguments{
\item{fun}{function name of an user-defined function}

\item{fn}{name of the function (string)}
}
\description{
Enhanced-Help Function: ehelp()
This function displays docstring style comments used as help liners for user
defined functions.
}
\examples{
myTestFn <- function(x,y,z,t=0) {
#'
#' This is just an example of a dummy fn
#'
#'
#' @email myemail@somewhere.org
#' @author author
#
#
#' @demo
#' @example myTestFn(x0,y0,z0)
}

ehelp(myTestFn)

}
\keyword{internal}
