% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehelp-utils.R
\name{print.ehelp}
\alias{print.ehelp}
\title{function associated to the ehelp object method}
\usage{
\method{print}{ehelp}(ehelp.obj, coloring = T)
}
\arguments{
\item{ehelp.obj}{ehelp object}

\item{coloring}{whether to use colors or not when displaying the help
keywords internal}
}
\description{
function associated to the ehelp object method
}
